Starting to connect to a remote kubernetes
This is about a million times harder than it needs to be when
dealing with running the tools on a different cluster to where
you want to deploy them to. There's /a lot/ more argument passing.

I wouldn't be as annoyed at this as I am, except that Google (in
their infinite wisdom) have provided a lot of really nice tools to
do one-stop-shop installs of their gcloud and kubectl CLI tools,
and essentially expect everyone to use them. I feel as though trying
to do things automatically (instead of interactively) you are out
on your own.

This is annoying because interactive feels like a toy. Great for
demos and a little bit of futzing here and there, but your mental
model is at best empty, at worst incomplete and wrong, making
automating of the tooling less simple than it should be.